Created StepTimer to handle the timer for countdown dials
